+++ src/cli/cli-application.ts
@@ -1,12 +1,13 @@
 import { CommandParser } from './command-parser.js';
 import { ICommand } from './commands/command.interface.js';
+import { COMMANDS } from './const/index.js';
 
 type TCommandCollection = Record<string, ICommand>;
 
 export class CLIApplication {
   private commands: TCommandCollection = {};
 
-  constructor(private readonly defaultCommand: string = '--help') {}
+  constructor(private readonly defaultCommand: string = COMMANDS.help) {}
 
   public registerCommands(commandList: ICommand[]): void {
     commandList.forEach((command) => {
