+++ src/cli/commands/help.command.ts
@@ -1,9 +1,10 @@
+import { COMMANDS } from '../const/index.js';
 import { ICommand } from './command.interface.js';
 import chalk from 'chalk';
 
 export class HelpCommand implements ICommand {
   public getName(): string {
-    return '--help';
+    return COMMANDS.help;
   }
 
   public async execute(..._parameters: string[]): Promise<void> {
@@ -12,10 +13,10 @@ export class HelpCommand implements ICommand {
         Пример:
             cli.js --<command> [--arguments]
         Команды:
-            ${chalk.gray('--version:')}                   # выводит номер версии
-            ${chalk.gray('--help:')}                      # печатает этот текст
-            ${chalk.gray('--import <path>:')}             # импортирует данные из TSV
-            ${chalk.gray('--generate <n> <path> <url>:')} # генерирует произвольное количество тестовых данных
+            ${chalk.gray(`${COMMANDS.version}:`)}                   # выводит номер версии
+            ${chalk.gray(`${COMMANDS.help}:`)}                      # печатает этот текст
+            ${chalk.gray(`${COMMANDS.import} <path>:`)}             # импортирует данные из TSV
+            ${chalk.gray(`${COMMANDS.generate} <n> <path> <url>:`)} # генерирует произвольное количество тестовых данных
     `);
   }
 }
