+++ src/cli/const/const.ts
@@ -1,0 +1,6 @@
+export const COMMANDS = {
+  help: '--help',
+  version: '--version',
+  import: '--import',
+  generate: '--generate',
+} as const;
