+++ src/shared/const/index.ts
@@ -1,0 +1,1 @@
+export * from './const.js';
