+++ src/shared/types/user-type.type.ts
@@ -1,8 +1,4 @@
+import { USER_TYPE } from '../const/index.js';
 import { TValueOf } from './value-of.type.js';
-
-export const USER_TYPE = {
-  simple: 'обычный',
-  pro: 'pro',
-} as const;
 
 export type TUserType = TValueOf<typeof USER_TYPE>;
